requirements.txt
aiodns==3.0.0
aiohttp==3.8.4      
aiosignal==1.3.1    
appdirs==1.4.4      
async-timeout==4.0.2
attrs==22.2.0       
Brotli==1.0.9       
certifi==2022.12.7  
cffi==1.15.1        
charset-normalizer==3.1.0
frozenlist==1.3.3
function-cooldowns==1.7.0
idna==3.4
multidict==6.0.4
nextcord==2.4.2
orjson==3.8.10
packaging==23.0
pycares==4.3.0
pycparser==2.21
PyExecJS==1.5.1
PyNaCl==1.4.0
pyparsing==3.0.9
pyTwistyScrambler==1.6
requests==2.28.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2